Robot Class Update
Removed unused using statement. Added comments.
Privatized strings and removed ones that were only used in methods from being global variables.
Renamed Methods.
Simplified logic in BuildPath so we are using a switch statement instead of a bunch of if else statements. Simplified the logic in DetermineFastestPath so that we only need to assign to finalPath once in each statement.